conditionally include novnc at build time.
