+++ src/component/utils/keyboards/novnc/browser.js
