+++ src/component/utils/keyboards/novnc/domkeytable.js
