+++ src/component/utils/keyboards/novnc/fixedkeys.js
