+++ src/component/utils/keyboards/novnc/keyboard.js
