+++ src/component/utils/keyboards/novnc/keysym.js
