+++ src/component/utils/keyboards/novnc/keysymdef.js
