+++ src/component/utils/keyboards/novnc/util.js
