+++ src/component/utils/keyboards/novnc/vkeys.js
